upgraded scallop and fixed other compilation problems of the archetype
