+++ pom.xml
@@ -32,7 +32,7 @@
     <description>Batch-updates metadata streams of datasets in a Fedora Commons repository</description>
     <inceptionYear>2016</inceptionYear>
     <properties>
-        <main-class>nl.knaw.dans.easy.module.Command</main-class>
+        <main-class>nl.knaw.dans.easy.umd.Command</main-class>
     </properties>
     <scm>
         <developerConnection>scm:git:https://github.com/DANS-KNAW/easy-update-metadata-dataset</developerConnection>
@@ -54,6 +54,7 @@
         <dependency>
             <groupId>org.rogach</groupId>
             <artifactId>scallop_2.11</artifactId>
+            <version>1.0.1</version>
         </dependency>
 		<dependency>
 			<groupId>commons-configuration</groupId>
